+++ changelog.md
@@ -1,37 +1,64 @@
 # 更新日志
 
-## [0.1.3]
+## [0.1.4]
+
+### 重要
+
+- 现在不再维护支持非 Windows 平台，当然欢迎提交 PR
+
 ### 新增
+
+### 优化
+
+- 优化了双击 `LeftAlt` 的时机判定
+- 使用 Windows API 来优化启动程序的逻辑，现在与 Windows 资源管理器双击的行为基本一致
+
+## [0.1.3]
+
+### 新增
+
 - 可以直接从文件管理器中拖拽程序来创建快捷方式
-- Windows平台支持使用exe图标作为快捷方式图片
+- Windows 平台支持使用 exe 图标作为快捷方式图片
 - 快捷方式现在支持参数
 - 快捷方式可以选择是否显示命令行
 - 快捷方式可以选择请求管理员权限运行
 - 现在可以控制是否激活双击 `LeftAlt` 呼出
+
 ### 优化
+
 - 将配置文件和缓存文件转移到文件夹中存储
 
 ## [0.1.2]
+
 ### 新增
+
 - 在编辑模式下，快捷方式可以拖拽排序
 - 可以为快捷方式添加标签，用于筛选
 - 快捷方式现在支持别名
 - 现在会尝试读取并修复旧版本的配置文件而不是覆盖
+
 ### 优化
-- 显著降低后台的CPU占用 (#5)
+
+- 显著降低后台的 CPU 占用 (#5)
 - 改进了搜索栏对大小写内容的支持
 
 ## [0.1.1]
+
 ### 新增
+
 - 工具箱现在可以在后台运行，并通过快捷键快速呼出
 - 添加编辑模式，可以新增/修改快捷方式
 - 搜索栏现在可以同时通过原文与拼音搜索
+
 ### 优化
 - 默认不再显示命令行 (#3)
 
 ## [0.1.0]
+
 ### 新增
+
 - 允许自主选择图片作为快捷方式图标
 - 可以右键配置快捷方式内容
+
 ### 优化
 - 现在会尝试使用系统中默认支持中文的字体
