+++ src/resources.rs
@@ -10,35 +10,7 @@ pub fn setup_custom_fonts(ctx: &egui::Context) {
     let mut fonts = egui::FontDefinitions::default();
 
     // 根据不同操作系统选择不同的字体路径
-    let font_path = if cfg!(target_os = "windows") {
-        // Windows系统下微软雅黑的默认位置
-        std::path::Path::new("C:/Windows/Fonts/msyh.ttc")
-    } else if cfg!(target_os = "linux") {
-        // Linux系统下常见的中文字体路径，尝试几个常见位置
-        let possible_paths = [
-            "/usr/share/fonts/noto/NotoSansCJK-Regular.ttc",         // Noto Sans CJK
-            "/usr/share/fonts/wenquanyi/wqy-microhei.ttc",           // 文泉驿微米黑
-            "/usr/share/fonts/wenquanyi/wqy-zenhei.ttc",             // 文泉驿正黑
-            "/usr/share/fonts/truetype/droid/DroidSansFallbackFull.ttf", // Droid Sans
-        ];
-        
-        // 查找第一个存在的字体文件
-        let mut found_path = std::path::Path::new("/usr/share/fonts"); // 默认路径
-        for path in possible_paths {
-            let full_path = std::path::Path::new(path);
-            if full_path.exists() {
-                found_path = full_path;
-                break;
-            }
-        }
-        found_path
-    } else if cfg!(target_os = "macos") {
-        // macOS系统下常见的中文字体
-        std::path::Path::new("/System/Library/Fonts/PingFang.ttc")  // 苹方字体
-    } else {
-        // 其他操作系统使用一个不太可能存在的路径，将回退到默认字体
-        std::path::Path::new("/non-existent-path")
-    };
+    let font_path = std::path::Path::new("C:/Windows/Fonts/msyh.ttc");
     
     if font_path.exists() {
         // 如果找到字体文件，从文件读取
