--- src/utils/mod.rs
@@ -1,3 +1,0 @@
-#[cfg(target_os = "windows")]
-pub mod windows_utils;
-
