+++ src/window/event.rs
@@ -5,15 +5,11 @@ pub enum UserEvent {
     Redraw(std::time::Duration),
     HideWindow,
     ShowWindow,
-    #[cfg(target_os = "windows")]
     Exit,
 
     // 托盘相关
-    #[cfg(target_os = "windows")]
     LeftClickTrayIcon,
-    #[cfg(target_os = "windows")]
     RightClickTrayIcon,
-    #[cfg(target_os = "windows")]
     ChangeDoubleAlt,
 
     // 文件相关
